Reintroduce TMADesc structure in copy.h and update its usage in atomic_add.cc. This change enhances the organization of TMA descriptor definitions and ensures consistent access across relevant files.
